requirements.txt
streamlit
groq
pydantic-ai
pydantic
python-dotenv